Power-Function
Write a function power() to raise a number m to power n. The function takes a double value for m and int value for n, and returns the result correctly. Use a default value of 2 for n to make the function to calculate sqaures when this value is omitted. Write a main that gets the values of m and n from the user to test the function.